% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{aemo_clear_cache}
\alias{aemo_clear_cache}
\title{Clear the aemo cache}
\usage{
aemo_clear_cache()
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Clear the aemo cache
}
\examples{
\donttest{
op <- options(aemo.cache_dir = tempdir())
aemo_clear_cache()
options(op)
}
}
\seealso{
Other configuration: 
\code{\link{aemo_cache_info}()},
\code{\link{aemo_throttle}()}
}
\concept{configuration}
