% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{aemo_regions}
\alias{aemo_regions}
\title{NEM regions}
\usage{
aemo_regions()
}
\value{
An \code{aemo_tbl} with columns \code{region}, \code{name}, \code{state},
\code{timezone}, \code{commenced}.
}
\description{
Returns a static table of the five NEM regions with metadata.
}
\examples{
aemo_regions()
}
\seealso{
Other reference: 
\code{\link{aemo_interconnectors}()},
\code{\link{aemo_units}()}
}
\concept{reference}
