% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemo.R
\name{clean_up_aemo_data_files}
\alias{clean_up_aemo_data_files}
\title{Remove AEMO CSV files}
\usage{
clean_up_aemo_data_files(path = ".")
}
\arguments{
\item{path}{a string containing the location of the AEMO (CSV) data files
that you wish to collate. These must have filenames of the form specified
above.}
}
\value{
no return value. However, AEMO files that are identified are removed.
}
\description{
This function removes AEMO data files that are on the \code{path}. These
are identified by having a filename of the form
\verb{[REGION][YYYY][MM].csv} (e.g. \code{"NSW201405.csv"}).
}
