% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_aemo_data}
\alias{get_aemo_data}
\title{Download data from AEMO website}
\usage{
get_aemo_data(regions, years, months)
}
\arguments{
\item{regions}{a case invariant character vector of abbreviated AEMO regions.}

\item{years}{a numeric vector of years in the form \code{YYYY}.}

\item{months}{a numeric vector of months in the form \code{MM}.}
}
\value{
data files whose number corresponds to the maximum lenght of one of
the parameters. The file names have the pattern: "[REGION][YYYY][MM].csv"
}
\description{
Download data from AEMO website
}

