% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aemo_tbl.R
\name{print.aemo_tbl}
\alias{print.aemo_tbl}
\title{Print an aemo_tbl}
\usage{
\method{print}{aemo_tbl}(x, ...)
}
\arguments{
\item{x}{An \code{aemo_tbl}.}

\item{...}{Passed to the next print method.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an aemo_tbl
}
\examples{
x <- data.frame(settlementdate = Sys.time(), region = "NSW1", rrp = 80)
x <- structure(x, aemo_title = "Demo", aemo_source = "http://nemweb.com.au",
               aemo_licence = "AEMO Copyright Permissions Notice",
               aemo_retrieved = Sys.time(),
               class = c("aemo_tbl", "data.frame"))
print(x)
}
