% Generated by roxygen2 (4.0.1): do not edit by hand
\name{aemo_regions}
\alias{aemo_regions}
\title{AEMO data set 1998-2014}
\format{\code{"REGION"} (factor), \code{"SETTLEMENTDATE"} (POSIXct),
\code{"TOTALDEMAND"} (numeric), \code{"RRP"} (numeric) and
\code{"PERIODTYPE"} (factor).}
\usage{
aemo_regions()
}
\description{
A sample data set containing trading interval price and demand data sourced
from the AEMO website on or about 16 June 2014. This was downloaded and
collated using \code{\link{collate_aemo_data}}.
}
\examples{
data(aemo)
head(aemo)
}
\references{
\href{http://www.aemo.com.au/Electricity/Data/Price-and-Demand/Aggregated-Price-and-Demand-Data-Files}{AEMO website}
}

